Refactored authentication code to its own app
