+++ authentication/apps.py
@@ -1,0 +1,7 @@
+"""Authentication Apps"""
+from django.apps import AppConfig
+
+
+class AuthenticationConfig(AppConfig):
+    """Authentication AppConfig"""
+    name = 'authentication'
