+++ authentication/constants.py
@@ -1,0 +1,4 @@
+"""Authentication constants"""
+
+AUTH_TYPE_REGISTER = 'register'
+AUTH_TYPE_LOGIN = 'login'
