+++ open_discussions/urls.py
@@ -17,26 +17,14 @@ from django.conf import settings
 from django.conf.urls import include, url
 from django.conf.urls.static import static
 from django.contrib import admin
-from django.contrib.auth import views as auth_views
 from rest_framework_jwt.views import refresh_jwt_token
 
-from open_discussions.views import (
-    index,
-    login,
-    register,
-    confirmation_sent,
-    jwt_login_complete,
-)
-
+from open_discussions.views import index
 
 urlpatterns = [
     url(r'^admin/', admin.site.urls),
     url(r'^status/', include('server_status.urls')),
-    url(r'^login/$', login, name='login'),
-    url(r'^login/micromasters/complete', jwt_login_complete, name='jwt-complete'),
-    url(r'^register/$', register, name='register'),
-    url(r'^register/confirmation_sent/$', confirmation_sent, name='confirmation-sent'),
-    url(r'^logout/$', auth_views.logout, name='logout'),
+    url(r'', include('authentication.urls')),
     url(r'', include('social_django.urls', namespace='social')),
     url(r'', include('channels.urls')),
     url(r'', include('profiles.urls')),
