+++ open_discussions/views_test.py
@@ -7,7 +7,7 @@ import pytest
 from django.urls import reverse
 from rest_framework_jwt.settings import api_settings
 
-from open_discussions.features import ANONYMOUS_ACCESS, EMAIL_AUTH
+from open_discussions.features import ANONYMOUS_ACCESS
 from profiles.utils import image_uri
 
 pytestmark = [
@@ -92,18 +92,6 @@ def test_webpack_url_jwt(
     }
 
 
-def test_login_jwt_complete(settings, client, user, jwt_token):
-    """Verify that the jwt-complete view invalidates the JWT auth cookie"""
-    client.cookies[api_settings.JWT_AUTH_COOKIE] = jwt_token
-    client.force_login(user)
-
-    response = client.get(reverse('jwt-complete'))
-
-    cookie = response.cookies['cookie_monster']
-    assert cookie['max-age'] == 0
-    assert cookie['domain'] == settings.OPEN_DISCUSSIONS_COOKIE_DOMAIN
-
-
 def test_webpack_url_anonymous(settings, client, mocker, authenticated_site):
     """Verify that webpack bundle src shows up in production"""
     settings.GA_TRACKING_ID = 'fake'
@@ -138,45 +126,3 @@ def test_webpack_url_anonymous(settings, client, mocker, authenticated_site):
         'is_authenticated': False,
         'allow_anonymous': 'access',
     }
-
-
-@pytest.mark.parametrize('is_enabled', [True, False])
-@pytest.mark.parametrize('is_anonymous', [True, False])
-def test_login_render(settings, client, user, is_enabled, is_anonymous):
-    """Test that login page renders or returns a 404"""
-    settings.FEATURES[EMAIL_AUTH] = is_enabled
-
-    if not is_anonymous:
-        client.force_login(user)
-
-    response = client.get(reverse('login'), follow=True)
-
-    if is_enabled:
-        assert response.status_code == 200
-        if not is_anonymous:
-            assert response.redirect_chain == [
-                ('/', 302),
-            ]
-    else:
-        assert response.status_code == 404
-
-
-@pytest.mark.parametrize('is_enabled', [True, False])
-@pytest.mark.parametrize('is_anonymous', [True, False])
-def test_register_render(settings, client, user, is_enabled, is_anonymous):
-    """Test that register page renders or returns a 404"""
-    settings.FEATURES[EMAIL_AUTH] = is_enabled
-
-    if not is_anonymous:
-        client.force_login(user)
-
-    response = client.get(reverse('register'), follow=True)
-
-    if is_enabled:
-        assert response.status_code == 200
-        if not is_anonymous:
-            assert response.redirect_chain == [
-                ('/', 302),
-            ]
-    else:
-        assert response.status_code == 404
